+++ src/main/java/com/iRecipeNew/iRecipeNew/service/RecipeServiceImpl.java
@@ -2,10 +2,12 @@ package com.iRecipeNew.iRecipeNew.service;
 
 
 import com.iRecipeNew.iRecipeNew.domain.Recipe;
+import com.iRecipeNew.iRecipeNew.errors.RecipeNotFoundError;
 import com.iRecipeNew.iRecipeNew.repository.RecipeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,30 +24,61 @@ public class RecipeServiceImpl implements RecipeService {
 
     }
 
+
     @Override
     public Optional<Recipe> getRecipeById(Long id) {
+
         return recipeRepository.findById(id);
     }
 
     @Override
-    public void updateRecipeById(Recipe recipe, Long id) {
+    public Recipe updateRecipeById(Long id, Recipe recipe) {
 
+        Recipe updatedRecipe = getRecipeById(id).get();
+
+
+        if(recipe.getId()!=null){
+            updatedRecipe.setId(recipe.getId());
+        }if(recipe.getCategory()!=null){
+            updatedRecipe.setCategory(recipe.getCategory());
+        }if(recipe.getComments()!=null){
+            updatedRecipe.setComments(recipe.getComments());
+        }if(recipe.getCookTimeInMin()!=null){
+            updatedRecipe.setCookTimeInMin(recipe.getCookTimeInMin());
+        }if(recipe.getCuisine()!=null){
+            updatedRecipe.setCuisine(recipe.getCuisine());
+        }if(recipe.getDifficulty()!=null){
+            updatedRecipe.setDifficulty(recipe.getDifficulty());
+        }if(recipe.getDirections()!=null){
+            updatedRecipe.setDirections(recipe.getDirections());
+        }if(recipe.getName()!=null){
+            updatedRecipe.setName(recipe.getName());
+        }if(recipe.getPrepTimeInMin()!=null){
+            updatedRecipe.setPrepTimeInMin(recipe.getPrepTimeInMin());
+        }if(recipe.getServings()!=null){
+            updatedRecipe.setServings(recipe.getServings());
+        }if(recipe.getUser()!=null){
+            updatedRecipe.setUser(recipe.getUser());
+        }
+
+       return this.recipeRepository.save(updatedRecipe);
+    }
+
+
+
+    @Override
+    public boolean deleteRecipeById(Long id) {
+
+        if (recipeRepository.existsById(id)) {
+            recipeRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public void replaceRecipeById(Recipe recipe, Long id) {
-
-    }
-
-    @Override
-    public void deleteRecipeById(Long id) {
-
-        recipeRepository.deleteById(id);
-
-    }
-
-    @Override
-    public void createRecipe(Recipe recipe) {
+    public void createRecipe(Recipe recipe)
+    {
         recipeRepository.save(recipe);
-    }
+        }
 }
